[organization] Move cell-related files to their own folder
- there's enough now that it's clogging up the `body/` folder and it makes sense to
