+++ src/components/datagrid/body/cell/index.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+export { EuiDataGridCell } from './data_grid_cell';
+
+export { Cell } from './data_grid_cell_wrapper';
+
+export {
+  DataGridCellPopoverContext,
+  useCellPopover,
+} from './data_grid_cell_popover';
