+++ src/components/datagrid/body/data_grid_body_virtualized.tsx
@@ -24,7 +24,7 @@ import {
 import { useResizeObserver } from '../../observer/resize_observer';
 import { useDataGridHeader } from './header';
 import { useDataGridFooter } from './footer';
-import { Cell } from './data_grid_cell_wrapper';
+import { Cell } from './cell';
 import {
   EuiDataGridBodyProps,
   DataGridWrapperRowsContentsShape,
